refactor(main): extract thanks, window, and core IPC modules from index
Thin index.ts by moving fetchThanksData, URL/window helpers, and misc IPC
handlers into focused modules with dependency injection, preserving behavior.
